make struct for URLSession Body
