+++ yumemicoding/ContentView.swift
@@ -11,9 +11,6 @@ struct ContentView: View {
     
     var body: some View {
         NavigationStack {
-           
-            // こいつが悪い？
-//            Text("\(Date().getYearMonthDay().year)")
             
             Spacer()
