+++ yumemicoding/yumemicodingApp.swift
@@ -6,7 +6,6 @@
 //
 
 import SwiftUI
-import SwiftData
 
 @main
 struct yumemicodingApp: App {
